pyproject.toml
[tool.poetry]
name = "app-vukwm-bag-delivery"
version = "0.1.0"
description = ""
authors = ["ejwillemse <ejwillemse@gmail.com>"]
readme = "README.md"
packages = [{include = "app_vukwm_bag_delivery"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.15.2"
watchdog = "^2.2.0"
pylint = "^2.15.8"
black = "^22.10.0"
openpyxl = "^3.0.10"
googlemaps = "^4.7.3"
toml = "^0.10.2"
pyxlsb = "^1.0.10"
XlsxWriter = "^3.0.3"
geopandas = "^0.12.1"
streamlit-aggrid = "^0.3.3"
hygese = "^0.0.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
